src/test/java/com/leetcode/dp/easy/IsSubsequence_392Test.java
package com.leetcode.dp.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class IsSubsequence_392Test {

  private final IsSubsequence_392 solution = new IsSubsequence_392();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments("abc", "ahbgdc", true), arguments("axc", "ahbgdc", false));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void isSubsequence(String s, String t, boolean expectedResult) {
    boolean actualResult = solution.isSubsequence(s, t);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
